Move core pipeline utilities into application layer
